pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = "Template for creating FastAPI projects"
authors = ["Pavel Vrukhin <vrukhin@protonmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_template"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.82.0"
pydantic = "^1.10.2"
uvicorn = "^0.18.3"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy = "^1.4.42"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
